Refactor App/Form details classes in integration tests
Refactor AppDetails and FormDetails to use a shared base class so that
it reduces code duplication
