+++ test_integration/lib/form_attribute.js
@@ -1,0 +1,7 @@
+module.exports = class FormAttribute {
+  constructor(value, elementId, isEditable=true) {
+    this.value = value;
+    this.elementId = elementId;
+    this.isEditable = isEditable;
+  }
+};
